Divide sidebar to app and vuestic sidebar, remove vuex functionality
